clean

+++ build.sh
@@ -3,6 +3,8 @@
 # Default values
 CMAKE_ARGS=""
 
+CLEAN=false
+
 # Find a valid macOS SDK and set it for CMake to fix a potential mismatch
 if [[ "$(uname)" == "Darwin" ]]; then
     SDK_PATH=$(xcrun --sdk macosx --show-sdk-path 2>/dev/null)
@@ -31,21 +33,25 @@ do
         CMAKE_ARGS="$CMAKE_ARGS -DCMAKE_CXX_CLANG_TIDY='clang-tidy;-fix-errors'"
         shift # Remove --tidy-fix from processing
         ;;
+        --clean)
+        CLEAN=true
+        shift
+        ;;
         *)
-        # Other arguments can be passed to cmake
+        CMAKE_ARGS="$CMAKE_ARGS $arg"
         ;;
     esac
 done
 
 # Clean build directory for a fresh configuration
-if [ -d "build" ]; then
+if $CLEAN && [ -d "build" ]; then
     echo "Removing existing build directory."
     rm -rf build
 fi
 
-echo "Configuring with: cmake -B build $CMAKE_ARGS"
+echo "Configuring with: cmake -B build -G Ninja $CMAKE_ARGS"
 
-# Run CMake configuration. This will generate compile_commands.json
+# Run CMake configuration.
 cmake -B build $CMAKE_ARGS
 
 # Run the build
